Remove all illegal characters from filename
Simply replace all illegal characters in the epub filename with a hyphen.

Illegal characters for NTFS and FAT are included. The period character is also
included to avoid filenames with a leading period. We aren't preventing
creating files named AUX or COM3, which are actually illegal on Windows too.

fixes #5
closes #8