Allow passing either story or chapter urls to the scraper
Extract the first number from the url and assume it's a chapter or story id. If
it's a valid story id, use it. If not, use the v2 info api to retrieve the
chapter info which contains the story id, and return that.

This uses a new wattpad api, the v2 info api. We can pass it a chapter id and
find out the story id from it.